Change walkthrough location defaults based on environment
If we set the WALKTHROUGH_LOCATIONS env var in the npm command then
it cannot be easily overridden. This change instead moves the default
values that were previously set in the npm commands into the server
itself.

If NODE_ENV is production, then the integr8ly repo will be used.
If NODE_ENV is not production, then the local directory will be used.

If WALKTHROUGH_LOCATIONS is defined, it will take priority.